add data file
new update

+++ APPS/SpeechApp/SpeechVideos/ViewModels/VideoListViewModel.cs
@@ -1,4 +1,4 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 using Microsoft.Practices.Prism.Commands;
 using Microsoft.Practices.Prism.Mvvm;
 using Microsoft.Practices.Prism.Regions;
@@ -11,6 +11,7 @@ using System.Threading;
 using System.Windows.Forms;
 using System.Windows.Data;
 using System.Windows.Input;
+using System.Text.RegularExpressions;
 using SpeechTTS.Model;
 using SpeechInfrastructure;
 
@@ -29,6 +30,10 @@ namespace SpeechVideos.ViewModels
         private readonly DelegateCommand<TextDocument> deleteTextCommand;
         private readonly DelegateCommand addNewCommand;
         private readonly ObservableCollection<TextDocument> textCollection;
+
+        private string readme = "";
+        private bool showIsChecked = false;
+        private bool notShowIsChecked = true;
 
         [ImportingConstructor]
         public VideoListViewModel(ITTService ttsService, IRegionManager regionManager)
@@ -58,6 +63,45 @@ namespace SpeechVideos.ViewModels
                         null);
                 },
                 null);
+
+            readme  = "课堂活动模块记录每节课或每个星期的教学内容，包括视频，句型，对话和生词。\n";
+            readme += "学生不仅能课堂跟老师学习，家里也可以反复练习，做到每星期的教学内容能100%消化。\n\n";
+            readme += "学生光凭课堂一点时间学习英语是远远不够的，课外必须有学习英语的环境，ECT的软件系统给学生\n";
+            readme += "提供良好的课外语言环境，只要有心每天坚持学习，不需要出国，也一定能全面学好英文。";
+        }
+
+        public bool ShowIsChecked
+        {
+            get
+            {
+                return this.showIsChecked;
+            }
+
+            set
+            {
+                this.SetProperty(ref this.showIsChecked, value);
+            }
+        }
+
+        public bool NotShowIsChecked
+        {
+            get
+            {
+                return this.notShowIsChecked;
+            }
+
+            set
+            {
+                this.SetProperty(ref this.notShowIsChecked, value);
+            }
+        }
+
+        public String Readme
+        {
+            get
+            {
+                return this.readme;
+            }
         }
 
         public ICollectionView Messages { get; private set; }
@@ -88,6 +132,9 @@ namespace SpeechVideos.ViewModels
 
         private void deleteFile(TextDocument document)
         {
+            int idx = ttsService.ClassIdx(document);
+            if (idx < ttsService.getNumClasses()) return;
+
             textCollection.Remove(document);
             this.ttsService.RemoveActivitiesDocument(document);
         }
@@ -102,15 +149,13 @@ namespace SpeechVideos.ViewModels
                 doc1 = new TextDocument();
 
             OpenFileDialog dialog = new OpenFileDialog();
-            dialog.Filter =
-               "TXT|*.video";
+            dialog.Filter = "TXT|*" + TTService.CLASS;
             dialog.InitialDirectory = "C:\\";
-            dialog.Title = "Select a text file";
+            dialog.Title = "Select a class file";
             if (dialog.ShowDialog() == DialogResult.OK)
             {
                 string fname = dialog.FileName;
                 doc.FileName = fname;
-                doc.Subject = Path.GetFileNameWithoutExtension(fname);
                 doc.StudentId = doc1.StudentId;
                 doc.From = doc1.From;
                 string[] lines = File.ReadAllLines(fname);
@@ -118,7 +163,7 @@ namespace SpeechVideos.ViewModels
                 {
                     if (line.Contains(TTService.TITLE_KEY))
                     {
-                        string[] col = line.Split(new char[] { TTService.SEP_CHAR });
+                        string[] col = Regex.Split(line, TTService.SEP_CHAR);
                         doc.Subject = col[1];
                         break;
                     }
