+++ APPS/SpeechWords/ViewModels/StoryListViewModel.cs
@@ -1,4 +1,4 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 using Microsoft.Practices.Prism.Commands;
 using Microsoft.Practices.Prism.Mvvm;
 using Microsoft.Practices.Prism.Regions;
@@ -11,6 +11,7 @@ using System.Threading;
 using System.Windows.Forms;
 using System.Windows.Data;
 using System.Windows.Input;
+using System.Text.RegularExpressions;
 using SpeechTTS.Model;
 using SpeechInfrastructure;
 
@@ -29,6 +30,10 @@ namespace SpeechWords.ViewModels
         private readonly DelegateCommand<TextDocument> deleteTextCommand;
         private readonly DelegateCommand addNewCommand;
         private readonly ObservableCollection<TextDocument> textCollection;
+
+        private string readme = "";
+        private bool showIsChecked = false;
+        private bool notShowIsChecked = true;
 
         [ImportingConstructor]
         public StoryListViewModel(ITTService ttsService, IRegionManager regionManager)
@@ -58,6 +63,10 @@ namespace SpeechWords.ViewModels
                         null);
                 },
                 null);
+
+            readme = "对于小年龄的孩子来说，看图识字，看图听故事和看图重复讲故事都是有效的学习方法，因为有视觉，听觉和口音的反复印记。\n\n";
+            readme += "卡通书模块除了提供课堂用的动画书外，学生也可以自己制作卡通书，配有自己写的简单英文，自己的画和自己的英文录音.\n";
+            readme += "学生可以选择用标准美国发音，或用自己的录音读自己写的作文。让学生感觉学英文可以自己编写卡通书，有一种成就感。";
         }
 
         public ICollectionView Messages { get; private set; }
@@ -77,6 +86,39 @@ namespace SpeechWords.ViewModels
             get { return this.addNewCommand; }
         }
 
+        public bool ShowIsChecked
+        {
+            get
+            {
+                return this.showIsChecked;
+            }
+
+            set
+            {
+                this.SetProperty(ref this.showIsChecked, value);
+            }
+        }
+
+        public bool NotShowIsChecked
+        {
+            get
+            {
+                return this.notShowIsChecked;
+            }
+
+            set
+            {
+                this.SetProperty(ref this.notShowIsChecked, value);
+            }
+        }
+
+        public String Readme
+        {
+            get
+            {
+                return this.readme;
+            }
+        }
 
         private void loadFile(TextDocument document)
         {
@@ -88,6 +130,9 @@ namespace SpeechWords.ViewModels
 
         private void deleteFile(TextDocument document)
         {
+            int idx = ttsService.CardIdx(document);
+            if (idx < ttsService.getNumCarts()) return;
+
             textCollection.Remove(document);
             this.ttsService.RemoveCardsDocument(document);
         }
@@ -102,15 +147,13 @@ namespace SpeechWords.ViewModels
                 doc1 = new TextDocument();
 
             OpenFileDialog dialog = new OpenFileDialog();
-            dialog.Filter =
-               "TXT|*.cartoon";
+            dialog.Filter = "TXT|*" + TTService.CARTOON;
             dialog.InitialDirectory = "C:\\";
-            dialog.Title = "Select a text file";
+            dialog.Title = "Select a cartoon file";
             if (dialog.ShowDialog() == DialogResult.OK)
             {
                 string fname = dialog.FileName;
                 doc.FileName = fname;
-                doc.Subject = Path.GetFileNameWithoutExtension(fname);
                 doc.StudentId = doc1.StudentId;
                 doc.From = doc1.From;
 
@@ -119,7 +162,7 @@ namespace SpeechWords.ViewModels
                 {
                     if (line.Contains(TTService.TITLE_KEY))
                     {
-                        string[] col = line.Split(new char[] { TTService.SEP_CHAR });
+                        string[] col = Regex.Split(line, TTService.SEP_CHAR);
                         doc.Subject = col[1];
                         break;
                     }
